FIX: [ 1189823 ] objc.protocolNamed doesn't find all protocols
The formal protocol list should be more complete.  A new
``objc.protocolsForProcess()`` enumerates over all mach
headers and returns all of the protocols defined in the expected
place.  This fixes the scenario where an application defines a
protocol (i.e. for plugins) but does not define any classes that
conform to that protocol.  In that case, the protocol would
reside in the mach header but would not be reachable by
iterating over all classes.