The custom wrappers for class methods have been updated, luckily this was a change in one function in pyobjc-api.m (and a simular change in alloc_hack.m).
Als update 'selector.isClassMethod', this is now True for all methods of a
meta class (mostly for limited backward compatibility).

The unittests no longer hang the interpreter, but not every tests passes at
the moment. Some of these are to be expected (if an ObjC class has a class
and instance method with the same name: MyClass.selector is now a class method
instead of an instance method), others are bugs.

I've also not looked at the code for modifying classes yet, that code requires
updates as well.

--HG--
branch : selector-cleanup