* Small tweaks to build_framework * Run_tests now actually does something useful
The run_test script only runs the 2.6 and 2.7 tests
for now, I once again ran into what seems like distribute and/or 2to3
related issues when testing using python 3.x.

The run_test script generates an HTML file with a report
of the build and test results. This is basicly a summary
with the number of problems for every project and some
coloring to make it easier to check the state of the project
at a glance.

--HG--
branch : pyobjc-ancient