The blocks runtime has an encoded signature for blocks, use that.
With a recent enough compiler (where 'recent enough' is still fairly
old) the compiler stores a type encoding for the block in the block
structure. I already had code to extract the signature from the block,
the signature is now actually used when there is no information about
the block result in the metadata system.