Avoid creating bytes objects when possible on Python 3.3
Starting at Python 3.3 the unicode object stores values
with the minimal byte width (UCS1, UCS2 or UCS4 depending
on the value of the string). Because of this ASCII strings,
such as selector names, can be accessed as 'const char*'
by directly accessing the internal buffer of the unicode
object.

This means we can avoid creating bytes objects for the name,
which should be slightly faster and use less memory. In both
cases this is a micro optimization.

NOTE: There might be more opportunities to apply this shortcut.

--HG--
branch : core-cleanup