Add an __signature__ property to objc.function and objc.selector objects
The __signature__ is an inspect.Signature object and currently contains
minimal information, the information can be enhanced later on by
changing the code in objc._callable_docstr.

In the long run this should make the output of pydoc nicer for PyObjC
callables, that requires a change to the Python stdlib though (because
pydoc currently does not use inspect.signature) and therefore won't
happen before Python 3.4.

--HG--
branch : core-cleanup