Move creation of proxies for Python objects from OC_* to a C function
This replaces a set of class methods on the OC_* classes
to a single C function. This makes the conversion easier
to understand.

It should also make it easier to further optimize the
conversion process (for example by caching how specific
types should be converted).

In the process:

* Add global variables from the OC_* files to objc.options,
  (and those variables are only used by the new function)

* Add, and document, 'registerSetType' and 'registerDateType'

--HG--
branch : core-cleanup