Suppport for large return values from blocks
Calling blocks with a large return value requires a slighly different
calling convention, add support for this to block_support.m.

Getting called doesn't require updates and works fine (as demonstrated by a
new testcase)

The value of aBlock.__block_signature__ now defaults to the signature
from the ObjC runtime (if that is present).

--HG--
branch : core-cleanup