Documentation tweaks
This fixes most of the dead links in the documentation.

Also adds a 'make linkcheck' target, and restructures
notes/framework-wrappers.rst to make it possible to use
the link checker for the document references.

notes/framework-wrappers.rst is getting close to the
limit w.r.t. maintaining it, but should be fine for now.